+++ frontend/src/components/outliner/WorkspaceHeader.tsx
@@ -6,9 +6,8 @@ import { Progress } from '../ui/progress';
 import { useOutlinerDocument } from '@/hooks/useOutlinerDocument';
 import {  useDocument } from './contexts';
 
-import InstructionsDrawer from './Instructions';
 import WorkSpaceHeaderMenu from './workspace/WorkSpaceHeaderMenu';
-import AIDetectionButton from './workspace/AIDetectionButton';
+// import AIDetectionButton from './workspace/AIDetectionButton';
 import ActionButton from './workspace/ActionButton';
 
 
@@ -47,8 +46,7 @@ export const WorkspaceHeader: React.FC<WorkspaceHeaderProps> = ({
           </div>
         </div>
         <div className="flex items-center gap-2">
-        <InstructionsDrawer/>
-        <AIDetectionButton/>
+        {/* <AIDetectionButton/> */}
         <WorkSpaceHeaderMenu/>
         <ActionButton/>
           <Button
